+++ src/Lane.cc
@@ -281,7 +281,7 @@ void LaneHeader::SetRightBoundary(const Marking &_boundary)
 }
 
 //////////////////////////////////////////////////
-unsigned int LaneHeader::NumCheckpoints() const
+size_t LaneHeader::NumCheckpoints() const
 {
   return this->dataPtr->checkpoints.size();
 }
@@ -366,7 +366,7 @@ bool LaneHeader::RemoveCheckpoint(const int _cpId)
 }
 
 //////////////////////////////////////////////////
-unsigned int LaneHeader::NumStops() const
+size_t LaneHeader::NumStops() const
 {
   return this->dataPtr->stops.size();
 }
@@ -416,7 +416,7 @@ bool LaneHeader::RemoveStop(const int _waypointId)
 }
 
 //////////////////////////////////////////////////
-unsigned int LaneHeader::NumExits() const
+size_t LaneHeader::NumExits() const
 {
   return this->dataPtr->exits.size();
 }
@@ -613,7 +613,7 @@ bool Lane::SetId(const int _id)
 }
 
 //////////////////////////////////////////////////
-unsigned int Lane::NumWaypoints() const
+size_t Lane::NumWaypoints() const
 {
   return this->dataPtr->waypoints.size();
 }
@@ -731,7 +731,7 @@ void Lane::SetRightBoundary(const Marking &_boundary)
 }
 
 //////////////////////////////////////////////////
-unsigned int Lane::NumCheckpoints() const
+size_t Lane::NumCheckpoints() const
 {
   return this->dataPtr->header.NumCheckpoints();
 }
@@ -773,7 +773,7 @@ bool Lane::RemoveCheckpoint(const int _cpId)
 }
 
 //////////////////////////////////////////////////
-unsigned int Lane::NumStops() const
+size_t Lane::NumStops() const
 {
   return this->dataPtr->header.NumStops();
 }
@@ -803,7 +803,7 @@ bool Lane::RemoveStop(const int _waypointId)
 }
 
 //////////////////////////////////////////////////
-unsigned int Lane::NumExits() const
+size_t Lane::NumExits() const
 {
   return this->dataPtr->header.NumExits();
 }
